Put Room class in its own file
